Split tests into different files
