+++ src/lib.rs
@@ -880,22 +880,10 @@ pub fn convert_halo2_circuit<
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tests::primitives::P128Pow5T3 as OrchardNullifier;
     use ark_pallas::Fq;
-    use ark_std::rand::rngs::OsRng;
     use ff::Field;
-    use folding_schemes::utils::vec::is_zero_vec;
-    use halo2_gadgets::poseidon::primitives::*;
-    use halo2_gadgets::poseidon::*;
-    use halo2_proofs::circuit::AssignedCell;
-    use halo2_proofs::circuit::Layouter;
-    use halo2_proofs::circuit::SimpleFloorPlanner;
-    use halo2_proofs::circuit::Value;
-    use halo2_proofs::dev::MockProver;
     use halo2_proofs::pasta::Fp;
-    use halo2_proofs::plonk::Error;
     use halo2_proofs::poly::Rotation;
-    use std::marker::PhantomData;
 
     // (ColumnA - 1) * ColumnA = 0
     // must be expanded into 2 monomials:
@@ -946,562 +934,5 @@ mod tests {
         ];
 
         assert_eq!(result, must_be);
-    }
-
-    #[test]
-    fn test_fibonacci_success() -> Result<(), Error> {
-        let instance_column: Vec<Fp> = vec![1.into(), 1.into(), 55.into()];
-
-        let k = 4;
-        let circuit = FibonacciCircuit(PhantomData);
-        let (ccs, z, _) =
-            convert_halo2_circuit::<_, _, ark_pallas::Fq>(k, &circuit, &[&instance_column])?;
-
-        let prover = MockProver::run(k, &circuit, vec![instance_column]).unwrap();
-        assert!(prover.verify().is_ok());
-        assert!(is_zero_vec(&ccs.eval_at_z(&z).unwrap()));
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_fibonacci_fail() -> Result<(), Error> {
-        let instance_column: Vec<Fp> = vec![1.into(), 1.into(), 54.into()];
-
-        let k = 4;
-        let circuit = FibonacciCircuit(PhantomData);
-        let (ccs, z, _) =
-            convert_halo2_circuit::<_, _, ark_pallas::Fq>(k, &circuit, &[&instance_column])?;
-
-        let prover = MockProver::run(k, &circuit, vec![instance_column]).unwrap();
-        assert!(prover.verify().is_err());
-        assert!(!is_zero_vec(&ccs.eval_at_z(&z).unwrap()));
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_poseidon_success() -> Result<(), Error> {
-        let message = [Fp::random(OsRng), Fp::random(OsRng)];
-        let output = halo2_gadgets::poseidon::primitives::Hash::<
-            _,
-            OrchardNullifier,
-            ConstantLength<2>,
-            3,
-            2,
-        >::init()
-        .hash(message);
-
-        let k = 6;
-        let circuit = HashCircuit::<OrchardNullifier, 3, 2, 2> {
-            message: Value::known(message),
-            _spec: PhantomData,
-        };
-        let (ccs, z, _) = convert_halo2_circuit::<_, _, ark_pallas::Fq>(k, &circuit, &[&[output]])?;
-
-        let prover = MockProver::run(k, &circuit, vec![vec![output]]).unwrap();
-        assert_eq!(prover.verify(), Ok(()));
-        assert!(is_zero_vec(&ccs.eval_at_z(&z).unwrap()));
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_poseidon_fail() -> Result<(), Error> {
-        let message = [Fp::random(OsRng), Fp::random(OsRng)];
-        let output = 0.into();
-
-        let k = 6;
-        let circuit = HashCircuit::<OrchardNullifier, 3, 2, 2> {
-            message: Value::known(message),
-            _spec: PhantomData,
-        };
-        let (ccs, z, _) = convert_halo2_circuit::<_, _, ark_pallas::Fq>(k, &circuit, &[&[output]])?;
-
-        let prover = MockProver::run(k, &circuit, vec![vec![output]]).unwrap();
-        assert!(prover.verify().is_err());
-        assert!(!is_zero_vec(&ccs.eval_at_z(&z).unwrap()));
-
-        Ok(())
-    }
-
-    // Taken from https://github.com/icemelon/halo2-examples/blob/master/src/fibonacci/example2.rs
-
-    #[derive(Debug, Clone)]
-    struct FibonacciConfig {
-        advice: Column<Advice>,
-        selector: Selector,
-        instance: Column<Instance>,
-    }
-
-    #[derive(Debug, Clone)]
-    struct FibonacciChip<F: Field> {
-        config: FibonacciConfig,
-        _marker: PhantomData<F>,
-    }
-
-    impl<F: Field> FibonacciChip<F> {
-        fn construct(config: FibonacciConfig) -> Self {
-            Self {
-                config,
-                _marker: PhantomData,
-            }
-        }
-
-        fn configure(
-            meta: &mut ConstraintSystem<F>,
-            advice: Column<Advice>,
-            instance: Column<Instance>,
-        ) -> FibonacciConfig {
-            let selector = meta.selector();
-
-            meta.enable_equality(advice);
-            meta.enable_equality(instance);
-
-            meta.create_gate("add", |meta| {
-                //
-                // advice | selector
-                //   a    |   s
-                //   b    |
-                //   c    |
-                //
-                let s = meta.query_selector(selector);
-                let a = meta.query_advice(advice, Rotation::cur());
-                let b = meta.query_advice(advice, Rotation::next());
-                let c = meta.query_advice(advice, Rotation(2));
-
-                vec![s * (a + b - c)]
-            });
-
-            FibonacciConfig {
-                advice,
-                selector,
-                instance,
-            }
-        }
-
-        fn assign(
-            &self,
-            mut layouter: impl Layouter<F>,
-            nrows: usize,
-        ) -> Result<AssignedCell<F, F>, Error> {
-            layouter.assign_region(
-                || "entire fibonacci table",
-                |mut region| {
-                    self.config.selector.enable(&mut region, 0)?;
-                    self.config.selector.enable(&mut region, 1)?;
-
-                    let mut a_cell = region.assign_advice_from_instance(
-                        || "1",
-                        self.config.instance,
-                        0,
-                        self.config.advice,
-                        0,
-                    )?;
-                    let mut b_cell = region.assign_advice_from_instance(
-                        || "1",
-                        self.config.instance,
-                        1,
-                        self.config.advice,
-                        1,
-                    )?;
-
-                    for row in 2..nrows {
-                        if row < nrows - 2 {
-                            self.config.selector.enable(&mut region, row)?;
-                        }
-
-                        let c_cell = region.assign_advice(
-                            || "advice",
-                            self.config.advice,
-                            row,
-                            || a_cell.value().copied() + b_cell.value(),
-                        )?;
-
-                        a_cell = b_cell;
-                        b_cell = c_cell;
-                    }
-
-                    Ok(b_cell)
-                },
-            )
-        }
-
-        fn expose_public(
-            &self,
-            mut layouter: impl Layouter<F>,
-            cell: AssignedCell<F, F>,
-            row: usize,
-        ) -> Result<(), Error> {
-            layouter.constrain_instance(cell.cell(), self.config.instance, row)
-        }
-    }
-
-    #[derive(Default)]
-    struct FibonacciCircuit<F>(PhantomData<F>);
-
-    impl<F: Field> Circuit<F> for FibonacciCircuit<F> {
-        type Config = FibonacciConfig;
-        type FloorPlanner = SimpleFloorPlanner;
-
-        fn without_witnesses(&self) -> Self {
-            Self::default()
-        }
-
-        fn configure(meta: &mut ConstraintSystem<F>) -> Self::Config {
-            let advice = meta.advice_column();
-            let instance = meta.instance_column();
-            FibonacciChip::configure(meta, advice, instance)
-        }
-
-        fn synthesize(
-            &self,
-            config: Self::Config,
-            mut layouter: impl Layouter<F>,
-        ) -> Result<(), Error> {
-            let chip = FibonacciChip::construct(config);
-
-            let out_cell = chip.assign(layouter.namespace(|| "entire table"), 10)?;
-
-            chip.expose_public(layouter.namespace(|| "out"), out_cell, 2)?;
-
-            Ok(())
-        }
-    }
-
-    #[derive(Clone)]
-    struct HashConfig<const WIDTH: usize, const RATE: usize> {
-        pow5: Pow5Config<Fp, WIDTH, RATE>,
-        instance_column: Column<Instance>,
-    }
-
-    // Taken from https://github.com/zcash/halo2/blob/halo2_proofs-0.3.0/halo2_gadgets/src/poseidon/pow5.rs#L719
-    struct HashCircuit<
-        S: Spec<Fp, WIDTH, RATE>,
-        const WIDTH: usize,
-        const RATE: usize,
-        const L: usize,
-    > {
-        message: Value<[Fp; L]>,
-        _spec: PhantomData<S>,
-    }
-
-    impl<S: Spec<Fp, WIDTH, RATE>, const WIDTH: usize, const RATE: usize, const L: usize>
-        Circuit<Fp> for HashCircuit<S, WIDTH, RATE, L>
-    {
-        type Config = HashConfig<WIDTH, RATE>;
-        type FloorPlanner = SimpleFloorPlanner;
-
-        fn without_witnesses(&self) -> Self {
-            Self {
-                message: Value::unknown(),
-                _spec: PhantomData,
-            }
-        }
-
-        fn configure(meta: &mut ConstraintSystem<Fp>) -> Self::Config {
-            let state = (0..WIDTH).map(|_| meta.advice_column()).collect::<Vec<_>>();
-            let partial_sbox = meta.advice_column();
-
-            let rc_a = (0..WIDTH).map(|_| meta.fixed_column()).collect::<Vec<_>>();
-            let rc_b = (0..WIDTH).map(|_| meta.fixed_column()).collect::<Vec<_>>();
-
-            meta.enable_constant(rc_b[0]);
-
-            let pow5 = Pow5Chip::configure::<S>(
-                meta,
-                state.try_into().unwrap(),
-                partial_sbox,
-                rc_a.try_into().unwrap(),
-                rc_b.try_into().unwrap(),
-            );
-
-            let instance_column = meta.instance_column();
-            meta.enable_equality(instance_column);
-
-            Self::Config {
-                pow5,
-                instance_column,
-            }
-        }
-
-        fn synthesize(
-            &self,
-            config: Self::Config,
-            mut layouter: impl Layouter<Fp>,
-        ) -> Result<(), Error> {
-            let chip = Pow5Chip::construct(config.pow5.clone());
-
-            let message = layouter.assign_region(
-                || "load message",
-                |mut region| {
-                    let message_word = |i: usize| {
-                        let value = self.message.map(|message_vals| message_vals[i]);
-                        region.assign_advice(
-                            || format!("load message_{}", i),
-                            config.pow5.state[i],
-                            0,
-                            || value,
-                        )
-                    };
-
-                    let message: Result<Vec<_>, Error> = (0..L).map(message_word).collect();
-                    Ok(message?.try_into().unwrap())
-                },
-            )?;
-
-            let hasher =
-                halo2_gadgets::poseidon::Hash::<_, _, S, ConstantLength<L>, WIDTH, RATE>::init(
-                    chip,
-                    layouter.namespace(|| "init"),
-                )?;
-            let output = hasher.hash(layouter.namespace(|| "hash"), message)?;
-
-            layouter.constrain_instance(output.cell(), config.instance_column, 0)
-        }
-    }
-
-    #[derive(Debug, Clone)]
-    struct RangeConfig {
-        table: TableColumn,
-        advice: Column<Advice>,
-    }
-
-    #[derive(Debug)]
-    struct RangeCircuit {
-        bytes: Vec<Fp>,
-    }
-
-    impl Circuit<Fp> for RangeCircuit {
-        type Config = RangeConfig;
-        type FloorPlanner = SimpleFloorPlanner;
-
-        fn without_witnesses(&self) -> Self {
-            Self { bytes: Vec::new() }
-        }
-
-        fn configure(meta: &mut ConstraintSystem<Fp>) -> Self::Config {
-            let advice = meta.advice_column();
-            let table = meta.lookup_table_column();
-
-            meta.lookup(|gate| {
-                let advice = gate.query_advice(advice, Rotation::cur());
-                vec![(advice, table)]
-            });
-
-            Self::Config { advice, table }
-        }
-
-        fn synthesize(
-            &self,
-            config: Self::Config,
-            mut layouter: impl Layouter<Fp>,
-        ) -> Result<(), Error> {
-            layouter.assign_region(
-                || "aa",
-                |mut region| {
-                    for (i, byte) in self.bytes.iter().enumerate() {
-                        region.assign_advice(|| "bb", config.advice, i, || Value::known(*byte))?;
-                    }
-                    Ok(())
-                },
-            )?;
-
-            layouter.assign_table(
-                || "range check",
-                |mut table| {
-                    for i in 0..(1 << 8) {
-                        table.assign_cell(
-                            || "a byte",
-                            config.table,
-                            i,
-                            || Value::known(Fp::from(i as u64)),
-                        )?;
-                    }
-                    Ok(())
-                },
-            )?;
-
-            Ok(())
-        }
-    }
-
-    // tests for the cases where the lookup input is a simple query
-    #[test]
-    fn test_range_success() -> Result<(), Error> {
-        let k = 9;
-        let circuit = RangeCircuit {
-            bytes: vec![1.into(), 2.into(), 3.into(), 255.into()],
-        };
-        let (ccs, z, lookups) = convert_halo2_circuit::<_, _, ark_pallas::Fq>(k, &circuit, &[])?;
-
-        let prover = MockProver::run(k, &circuit, vec![]).unwrap();
-        assert_eq!(prover.verify(), Ok(()));
-        assert!(is_zero_vec(&ccs.eval_at_z(&z).unwrap()));
-
-        let is_lookup_satisfied = lookups.into_iter().all(|(z_indices, table)| {
-            z_indices
-                .into_iter()
-                .map(|z_index| z[z_index])
-                .collect::<HashSet<_>>()
-                .is_subset(&table)
-        });
-        assert!(is_lookup_satisfied);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_range_failure() -> Result<(), Error> {
-        let k = 9;
-        let circuit = RangeCircuit {
-            bytes: vec![1.into(), 2.into(), 3.into(), 256.into()],
-        };
-        let (ccs, z, lookups) = convert_halo2_circuit::<_, _, ark_pallas::Fq>(k, &circuit, &[])?;
-
-        let prover = MockProver::run(k, &circuit, vec![]).unwrap();
-        assert!(prover.verify().is_err());
-
-        let is_lookup_satisfied = lookups.into_iter().all(|(z_indices, table)| {
-            z_indices
-                .into_iter()
-                .map(|z_index| z[z_index])
-                .collect::<HashSet<_>>()
-                .is_subset(&table)
-        });
-        assert!(!is_lookup_satisfied);
-
-        Ok(())
-    }
-
-    #[derive(Debug)]
-    struct LessThanCircuit {
-        less_than_200: Vec<Fp>,
-    }
-
-    #[derive(Debug, Clone)]
-    struct LessThanConfig {
-        table: TableColumn,
-        advice: Column<Advice>,
-        is_enabled: Column<Fixed>,
-    }
-
-    impl Circuit<Fp> for LessThanCircuit {
-        type Config = LessThanConfig;
-        type FloorPlanner = SimpleFloorPlanner;
-
-        fn without_witnesses(&self) -> Self {
-            Self {
-                less_than_200: Vec::new(),
-            }
-        }
-
-        fn configure(meta: &mut ConstraintSystem<Fp>) -> Self::Config {
-            let advice = meta.advice_column();
-            let table = meta.lookup_table_column();
-            let is_enabled = meta.fixed_column();
-
-            meta.lookup(|gate| {
-                let advice = gate.query_advice(advice, Rotation::cur());
-                let is_enabled = gate.query_fixed(is_enabled);
-                vec![(
-                    is_enabled * (advice + Expression::Constant(Fp::from(56))),
-                    table,
-                )]
-            });
-
-            Self::Config {
-                advice,
-                table,
-                is_enabled,
-            }
-        }
-
-        fn synthesize(
-            &self,
-            config: Self::Config,
-            mut layouter: impl Layouter<Fp>,
-        ) -> Result<(), Error> {
-            layouter.assign_region(
-                || "aa",
-                |mut region| {
-                    for (i, byte) in self.less_than_200.iter().enumerate() {
-                        region.assign_fixed(
-                            || "cc",
-                            config.is_enabled,
-                            i,
-                            || Value::known(Fp::ONE),
-                        )?;
-                        region.assign_advice(|| "bb", config.advice, i, || Value::known(*byte))?;
-                    }
-                    Ok(())
-                },
-            )?;
-
-            layouter.assign_table(
-                || "range check",
-                |mut table| {
-                    for i in 0..(1 << 8) {
-                        table.assign_cell(
-                            || "a byte",
-                            config.table,
-                            i,
-                            || Value::known(Fp::from(i as u64)),
-                        )?;
-                    }
-                    Ok(())
-                },
-            )?;
-
-            Ok(())
-        }
-    }
-
-    // tests for the cases where the lookup input is a complex Expression
-    #[test]
-    fn test_less_than_success() -> Result<(), Error> {
-        let k = 9;
-        let circuit = LessThanCircuit {
-            less_than_200: vec![1.into(), 2.into(), 3.into(), 199.into()],
-        };
-        let (ccs, z, lookups) = convert_halo2_circuit::<_, _, ark_pallas::Fq>(k, &circuit, &[])?;
-
-        let prover = MockProver::run(k, &circuit, vec![]).unwrap();
-        assert_eq!(prover.verify(), Ok(()));
-        assert!(is_zero_vec(&ccs.eval_at_z(&z).unwrap()));
-
-        let is_lookup_satisfied = lookups.into_iter().all(|(z_indices, table)| {
-            z_indices
-                .into_iter()
-                .map(|z_index| z[z_index])
-                .collect::<HashSet<_>>()
-                .is_subset(&table)
-        });
-        assert!(is_lookup_satisfied);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_less_than_failure() -> Result<(), Error> {
-        let k = 9;
-        let circuit = LessThanCircuit {
-            less_than_200: vec![1.into(), 2.into(), 3.into(), 200.into()],
-        };
-        let (_, z, lookups) = convert_halo2_circuit::<_, _, ark_pallas::Fq>(k, &circuit, &[])?;
-
-        let prover = MockProver::run(k, &circuit, vec![]).unwrap();
-        assert!(prover.verify().is_err());
-
-        let is_lookup_satisfied = lookups.into_iter().all(|(z_indices, table)| {
-            z_indices
-                .into_iter()
-                .map(|z_index| z[z_index])
-                .collect::<HashSet<_>>()
-                .is_subset(&table)
-        });
-        assert!(!is_lookup_satisfied);
-
-        Ok(())
     }
 }
